fix(defender): Name check change and verify log agents installed
