+++ prowler/providers/azure/services/defender/defender_ensure_system_updates_are_applied/defender_ensure_system_updates_are_applied.py
@@ -2,7 +2,7 @@ from prowler.lib.check.models import Check, Check_Report_Azure
 from prowler.providers.azure.services.defender.defender_client import defender_client
 
 
-class defender_ensure_apply_system_updates_is_completed(Check):
+class defender_ensure_system_updates_are_applied(Check):
     def execute(self) -> Check_Report_Azure:
         findings = []
 
@@ -10,9 +10,10 @@ class defender_ensure_apply_system_updates_is_completed(Check):
             subscription_name,
             assessments,
         ) in defender_client.assessments.items():
-
             if (
-                "Machines should be configured to periodically check for missing system updates"
+                "Log Analytics agent should be installed on virtual machines"
+                in assessments
+                and "Machines should be configured to periodically check for missing system updates"
                 in assessments
                 and "System updates should be installed on your machines" in assessments
             ):
@@ -20,25 +21,29 @@ class defender_ensure_apply_system_updates_is_completed(Check):
                 report.status = "PASS"
                 report.subscription = subscription_name
                 report.resource_name = assessments[
-                    "Machines should be configured to periodically check for missing system updates"
+                    "System updates should be installed on your machines"
                 ].resource_name
                 report.resource_id = assessments[
-                    "Machines should be configured to periodically check for missing system updates"
+                    "System updates should be installed on your machines"
                 ].resource_id
-                report.status_extended = f"Apply system updates assessment is HEALTHY for subscription {subscription_name}."
+                report.status_extended = f"System updates are applied for all the VMs in the subscription {subscription_name}."
 
                 if (
                     assessments[
+                        "Log Analytics agent should be installed on virtual machines"
+                    ].status
+                    == "Unhealthy"
+                    or assessments[
                         "Machines should be configured to periodically check for missing system updates"
                     ].status
-                    != "Healthy"
+                    == "Unhealthy"
                     or assessments[
                         "System updates should be installed on your machines"
                     ].status
-                    != "Healthy"
+                    == "Unhealthy"
                 ):
                     report.status = "FAIL"
-                    report.status_extended = f"Apply system updates assessment is UNHEALTHY for subscription {subscription_name}."
+                    report.status_extended = f"System updates are not applied for all the VMs in the subscription {subscription_name}."
 
                 findings.append(report)
